Using multiple prices for the same product
